+++ packages/common/src/schemas/documents/default-context-schema.ts
@@ -16,6 +16,7 @@ const entitySchema = Type.Object(
 );
 
 export const defaultContextSchema = Type.Object({
+  aml: Type.Optional(Type.Unknown()),
   entity: Type.Union([
     Type.Composite([entitySchema, Type.Object({ id: Type.String() })]),
     Type.Composite([entitySchema, Type.Object({ ballerineEntityId: Type.String() })]),
@@ -75,18 +76,19 @@ export const defaultContextSchema = Type.Object({
                   Type.String(),
                   Type.String({
                     enum: [
-                      'Wrong document',
-                      'Fake document',
+                      'Wrong category',
                       'Spam',
                       'Ownership mismatch - Name',
                       'Ownership mismatch - National ID',
-                      'Unknown document type',
                       'Bad image quality',
                       'Missing page',
                       'Invalid document',
                       'Expired document',
-                      'Unreadable document',
+                      'Password protected',
                       'Blurry image',
+                      'Short statement period',
+                      'Statement out of range',
+                      'Outside restricted area',
                       'Other',
                     ],
                   }),