+++ services/workflows-service/scripts/workflows/ui-definition/kyb-with-associated-companies/definition/compose-kyb-with-associated-companies-definition.ts
@@ -22,12 +22,12 @@ export const composeKybWithAssociatedCompaniesDefinition = ({
     definition: {
       id: `${definitionId}_v1`,
       predictableActionArguments: true,
-      initial: 'idle',
+      initial: 'collection_flow',
       context: {
         documents: [],
       },
       states: {
-        idle: {
+        collection_flow: {
           tags: [StateTag.COLLECTION_FLOW],
           on: {
             COLLECTION_FLOW_FINISHED: [{ target: 'create_kyc_workflows' }],
@@ -238,6 +238,7 @@ export const composeKybWithAssociatedCompaniesDefinition = ({
           deliverEvent: 'KYC_REVISION',
         },
       ],
+      isCaseOverviewEnabled: true,
     },
     contextSchema: {
       type: 'json-schema',