+++ services/workflows-service/src/common/filters/prisma-client-validation-filter/utils/parsers/invalid-argument-parser/invalid-argument.parser.ts
@@ -11,6 +11,7 @@ export class InvalidArgumentParser extends IParser {
 
   parse(): IParserResult {
     const { message } = this;
+
     if (!message) return {};
 
     return this.execPattern(this.pattern, (result, match) => {