+++ services/workflows-service/src/health/health.controller.ts
@@ -15,9 +15,11 @@ export class HealthController {
   @Get(ROUTES.READY)
   async healthReady(@Res() response: Response): Promise<Response<void>> {
     const dbConnection = await this.healthService.isDbReady();
+
     if (!dbConnection) {
       return response.status(HttpStatus.NOT_FOUND).send();
     }
+
     return response.status(HttpStatus.NO_CONTENT).send();
   }
 }