+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/workflows-dashboard
 
+## 0.1.14
+
+### Patch Changes
+
+- version bump
+
+## 0.1.13
+
+### Patch Changes
+
+- Ver bump
+
 ## 0.1.12
 
 ### Patch Changes