+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/blocks
 
+## 0.1.28
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.46
+
+## 0.1.27
+
+### Patch Changes
+
+- Ver bump
+- Updated dependencies
+  - @ballerine/common@0.7.43
+
 ## 0.1.26
 
 ### Patch Changes