+++ packages/common/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ballerine/common
 
+## 0.7.46
+
+### Patch Changes
+
+- version bump
+
+## 0.7.45
+
+### Patch Changes
+
+- Updated common enums
+
+## 0.7.44
+
+### Patch Changes
+
+- Fixes for concurrency fixes
+
+## 0.7.43
+
+### Patch Changes
+
+- Ver bump
+
+## 0.7.42
+
+### Patch Changes
+
+- updated document schema validation and text
+
 ## 0.7.41
 
 ### Patch Changes